Fixed a bug that would prevent big numbers from being queued correctly

+++ huffman.h
@@ -2,6 +2,7 @@
 # define _HUFFMAN_H
 # include <stdint.h>
 # include <stdbool.h>
+# include <stdio.h>
 
 # ifndef NIL
 # define NIL (void *) 0
@@ -37,7 +38,7 @@ void delTree(treeNode *t);
 
 static inline void delNode(treeNode *h) { free(h); return; }
 
-static inline int8_t compare(treeNode *l, treeNode *r)
+static inline int32_t compare(treeNode *l, treeNode *r)
 {
   return l->count - r->count; // l < r if negative, 1 = r if 0, ...
 }
